+++ deprecated_functions/connect_to_phone.py
