+++ deprecated_functions/send_whatsapp_message.py
@@ -1,9 +1,9 @@
 import os
 import subprocess
 import time
-from messaging_functions.connect_to_phone import connect_adb_wirelessly
-from messaging_functions.connect_to_phone import connect_adb_via_usb
-from messaging_functions.unlock_phone import unlock_device
+from deprecated_functions.connect_to_phone import connect_adb_wirelessly
+from deprecated_functions.connect_to_phone import connect_adb_via_usb
+from deprecated_functions.unlock_phone import unlock_device
 
 
 def send_whatsapp_message(phone_number, message, device_id=None, connectivity_method=os.getenv('CONNECTIVITY_METHOD')):
